Rename BPRecord to BP; Records are activities now and BPRecord is actually BP
