+++ frontend/src/components/manage/Blocks/BusinessPlans/BP/BPDetails.tsx
